Use short for machine identifiers

+++ backend/src/controlmat.Application/Common/Dto/ActiveWashDto.cs
@@ -7,7 +7,7 @@ namespace Controlmat.Application.Common.Dto;
 public class ActiveWashDto
 {
     public long WashingId { get; set; }
-    public int MachineId { get; set; }
+    public short MachineId { get; set; }
     public string MachineName { get; set; } = string.Empty;
     public DateTime StartDate { get; set; }
     public string StartUserName { get; set; } = string.Empty;
