+++ backend/src/controlmat.Application/Common/Dto/WashingResponseDto.cs
@@ -8,7 +8,7 @@ public class WashingResponseDto
 {
     public long WashingId { get; set; }
 
-    public int MachineId { get; set; }
+    public short MachineId { get; set; }
     public string MachineName { get; set; } = string.Empty;
     public int StartUserId { get; set; }
     public string StartUserName { get; set; } = string.Empty;
